+++ ricaun.RevitTest.Tests/TestsDebugger.cs
@@ -1,6 +1,12 @@
 ﻿using NUnit.Framework;
 using System;
 using System.Diagnostics;
+using System.Reflection;
+
+//[assembly: AssemblyMetadata("NUnit.RevitVersion", "2022")]
+[assembly: AssemblyMetadata("NUnit.RevitOpen", "false")]
+[assembly: AssemblyMetadata("NUnit.RevitClose", "false")]
+[assembly: AssemblyMetadata("NUnit.Verbosity", "5")]
 
 namespace ricaun.RevitTest.Tests
 {